Manage configuration settings of lists
If we are managing mailname lists then we should be able to configure
them as well. This formula won't detect changes that are made to list
configuration settings outside of salt (ie. via the web interface), so
changes will only be made and/or enforced when changes are made to the
pillar data.