Fix conditional update of list configs
The `onlyif` argument does not work with `cmd.wait` and `cmd.wait` seems
like it will be deprecated in favour of using `cmd.run` with the
`onchanges` requisite [1].

It also seems to make sense to update the config if the `newlist`
command has been run, otherwise there is a chance that the command may
fail but the config file is created. If the command is later run
successfully but the config file doesn't change, then the config update
command would not run.

[1] https://github.com/saltstack/salt/issues/17177#issuecomment-69621331